+++ src/sys/dev/pci/if_lge.c
@@ -602,7 +602,7 @@ void lge_attach(parent, self, aux)
 			   sizeof(struct lge_list_data), &kva,
 			   BUS_DMA_NOWAIT)) {
 		printf("%s: can't map dma buffers (%d bytes)\n",
-		       sc->sc_dv.dv_xname, sizeof(struct lge_list_data));
+		       sc->sc_dv.dv_xname, (int)sizeof(struct lge_list_data));
 		bus_dmamem_free(sc->sc_dmatag, &seg, rseg);
 		goto fail;
 	}
@@ -835,7 +835,7 @@ int lge_alloc_jumbo_mem(sc)
 	}
 	if (bus_dmamem_map(sc->sc_dmatag, &seg, rseg, LGE_JMEM, &kva,
 			   BUS_DMA_NOWAIT)) {
-		printf("%s: can't map dma buffers (%d bytes)\n",
+		printf("%s: can't map dma buffers (%ld bytes)\n",
 		       sc->sc_dv.dv_xname, LGE_JMEM);
 		bus_dmamem_free(sc->sc_dmatag, &seg, rseg);
 		return (ENOBUFS);
